Makefile: split python_texinfo target
python_texinfo is now a .PHONY target that depends on the python.texi
file, which was what the target produced before.

python.texi has a rule of its own that does any post-processing of the
texi file and copies it to the top level.

A new target python_info runs sphinx-build on the Python docs.